Backport: Improve server tick loop and implement tps api
This patch was backported from Paper 1.14.4 to Project Poseidon. The original author of the patch in Paper is Aikar.

This patch provides a more stable server tick loop that is not affected by changes in the system time. In addition, it is intended to keep the server at 20 TPS unless it can not realistically do so. This patch also allows Project Poseidon to keep track of server TPS, which can be viewed with the /tps command.